frontend/starter_app.py
import streamlit as st
import datetime

# --- Branding & Layout ---
st.set_page_config(page_title="DevAgent Swarm Dashboard", layout="wide", page_icon="🛡️")
st.image("https://streamlit.io/images/brand/streamlit-logo-secondary-colormark-darktext.png", width=180)
st.title("DevAgent Swarm: AI Code Review & Auto-Fix Dashboard 🛡️")

# --- Sidebar: Authentication & Repo Selection ---
st.sidebar.header("Authentication & Repo Selection")
github_token = st.sidebar.text_input("GitHub Token", type="password")
repo_url = st.sidebar.text_input("GitHub Repo URL", value="23f3000340/covid-xray-classifier")
if st.sidebar.button("Connect & Analyze"):
    st.session_state["connected"] = True
    st.session_state["last_scan"] = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
    # TODO: Trigger backend analysis here

# --- Dashboard Overview ---
st.subheader("Dashboard Overview")
col1, col2, col3 = st.columns(3)
col1.metric("Files Analyzed", "3")
col2.metric("Vulnerabilities Found", "7")
col3.metric("Last Scan", st.session_state.get("last_scan", "Never"))

# --- Progress Bar/Spinner ---
if st.session_state.get("connected"):
    with st.spinner("Analyzing repository..."):
        # TODO: Replace with actual backend call
        st.success("Analysis complete!")

# --- Detailed Findings Table ---
st.subheader("Findings")
findings = [
    {"File": "valub.py", "Line": 19, "Severity": "CRITICAL", "Description": "Hardcoded credentials", "Recommended Fix": "Use env variable"},
    {"File": "adi_valub.py", "Line": 3, "Severity": "HIGH", "Description": "Race condition", "Recommended Fix": "Use lock"},
    {"File": "app.py", "Line": 79, "Severity": "MEDIUM", "Description": "Potential XSS", "Recommended Fix": "Sanitize input"}
]
severity_colors = {"CRITICAL": "#ff4b4b", "HIGH": "#ff9800", "MEDIUM": "#ffd600", "LOW": "#8bc34a"}
for f in findings:
    with st.expander(f"{f['Severity']} | {f['File']} (Line {f['Line']})"):
        st.write(f"**Description:** {f['Description']}")
        st.write(f"**Recommended Fix:** {f['Recommended Fix']}")
        st.code("# Show code snippet here\n...")
        if st.button(f"Apply & Commit Fix for {f['File']} Line {f['Line']}", key=f"commit_{f['File']}_{f['Line']}"):
            # TODO: Backend commit logic
            st.success("Fix applied and committed to GitHub!")

# --- Visualization & Insights ---
st.subheader("Insights")
st.write("Pie/bar chart of issue types/severity (demo)")
# TODO: Add chart with matplotlib/plotly

# --- Notifications & Feedback ---
st.subheader("Feedback & Report")
if st.button("Download Report (Markdown)"):
    st.info("Report downloaded (demo)")
with st.form("feedback_form"):
    feedback = st.text_area("Your feedback")
    submitted = st.form_submit_button("Submit Feedback")
    if submitted:
        st.success("Thank you for your feedback!")

# --- Professional Touches ---
st.markdown("---")
st.markdown("_Built by DevAgent Swarm for hackathons & beyond!_")
